+++ util.h
@@ -40,7 +40,7 @@ void free_array(mag_array*);
 
 mag_array make_adj_list();
 
-int add_edge(mag_array *adj_list, int source, int sink);
+int add_edge(mag_array *adj_list, long source, long sink);
 
 void destroy_adj_list(mag_array *adj_list);
